exa: factor out 2d / compositing to separate files
To ease working with the code a tad.
